+++ CHANGES.md
@@ -11,3 +11,8 @@ CHANGES
 1.2.0
 
 - support align attribute in ```<text>``` node with ImageDrawPrint
+
+1.3.0
+
+- EscPosPrint auto convert zh-hant to zh-hans when encode with gbk or gb2312.
+- EscPosPrint ignore encode error.